Add Get Submissions Endpoint
[GET] tasks/<taskID>/submissions
query:
   userID=<integer>
   submission=<boolean>

If a user ID is specified in the query, then return the submission of that specific user.
If not, return all the submissions of the users assigned with that task.
Submission parameter decides if the solution is required to be returned or not.
The default behavior is false. Thus, only the grade and the user information are returned.